refactor: split submissions mixin
Move views and JSON handlers to new files, keep base utils
